fix(retry): add minimum delay between retry attempts and update default value

+++ crates/forge_domain/src/orch.rs
@@ -600,7 +600,7 @@ impl<S: AgentService> Orchestrator<S> {
                 .await?;
 
             // Set estimated tokens
-            usage.estimated_tokens = estimate_token_count(context.to_text().len() as usize) as u64;
+            usage.estimated_tokens = estimate_token_count(context.to_text().len()) as u64;
 
             // Send the usage information if available
